requirements.txt
click>=8.1.7
rich>=13.7.0
google-generativeai>=0.3.2
python-dotenv>=1.0.0
requests>=2.31.0
PyGithub>=2.1.1
groq>=0.4.0

